Shell script to report RSSI values of devices connected to a dd-wrt router
Initial commit - I wrote this for my personal setup and it works on dd-wrt ARM v3.0-r28600M. Add the script to your router and run at boot to report values to your find-lf server. I find that the values are more accurate than any of my pi's running tshark. It simply adds the router as an additional node in the cluster.
I appologize for not making it more robust. I'm very new to shell text manipulation/scripting and wrote this to practice - figured I'd share. Hopefully someone can clean up, standardize, and integrate with the cluster config commands.